Multiplies two positive numbers

+++ 0x0C-more_malloc_free/101-mul.c
@@ -29,7 +29,9 @@ int main(int argc, char *argv[])
 			}
 		}
 	}
-	mul = atol(argv[1]) * atol(argv[2]);
+
+	mul = atoi(argv[1]) * atoi(argv[2]);
 	printf("%lu\n", mul);
+
 	return (0);
 }
